@@ -2,6 +2,7 @@
   "name": "native",
   "version": "1.0.0",
   "main": "expo-router/entry",
+  "private": true,
   "scripts": {
     "start": "expo start",
     "start:tunnel": "expo start --tunnel",
@@ -20,7 +21,7 @@
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "e2e:test": "bash -c \"maestro test ../../packages/.maestro/$1\"",
-    "lint": "expo lint  --ignore-pattern '**/.next/**' --ignore-pattern '.history/**' ",
+    "lint": "eslint . --ignore-pattern '**/.next/**' --ignore-pattern '.history/**'",
     "lint-fix": "eslint .  --ignore-pattern '**/.next/**' --ignore-pattern '.history/**'  --fix",
     "test:tsc": " tsc",
     "pre-commit": "lint-staged",
@@ -32,10 +33,12 @@
     "preset": "jest-expo"
   },
   "dependencies": {
+    "@expo-google-fonts/dev": "^0.4.40",
     "@expo/vector-icons": "^15.0.2",
     "@radix-ui/react-compose-refs": "^1.1.2",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-native/virtualized-lists": "^0.77.1",
+    "@react-navigation/bottom-tabs": "^7.8.5",
     "@react-navigation/native": "^7.0.14",
     "@shopify/flash-list": "2.0.2",
     "add": "^2.0.6",
@@ -54,8 +57,6 @@
     "expo-web-browser": "~15.0.8",
     "install": "^0.13.0",
     "jotai": "^2.9.0",
-    "metro-react-native-babel-transformer": "^0.77.0",
-    "prettier": "^3.3.2",
     "react": "19.1.0",
     "react-compiler-runtime": "^19.1.0-rc.2",
     "react-dom": "19.1.0",
@@ -66,22 +67,22 @@
     "react-native-safe-area-context": "~5.6.0",
     "react-native-screens": "~4.16.0",
     "react-native-svg": "15.12.1",
-    "react-native-svg-transformer": "^1.5.0",
     "react-native-web": "^0.21.0",
     "react-native-worklets": "0.5.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@babel/eslint-parser": "^7.5.4",
+    "@treetracker/config": "0.0.1",
     "@types/chai": "^5.2.2",
     "@types/jest": "^29.5.12",
     "@types/mocha": "^10.0.10",
-    "@types/node": "^24.9.1",
+    "@types/node": "^24.10.1",
     "@types/react": "~19.1.10",
     "@types/react-test-renderer": "^19.0.10",
     "@types/webdriverio": "^5.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.46.2",
-    "@typescript-eslint/parser": "^7.16.0",
+    "@typescript-eslint/eslint-plugin": "^8.46.4",
+    "@typescript-eslint/parser": "^8.46.4",
     "@wdio/appium-service": "^9.20.0",
     "@wdio/cli": "^9.20.0",
     "@wdio/cucumber-framework": "^9.20.0",
@@ -89,22 +90,21 @@
     "@wdio/mocha-framework": "^9.20.0",
     "@wdio/spec-reporter": "^9.20.0",
     "@wdio/types": "^9.20.0",
+    "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-react-compiler": "^19.1.0-rc.2",
     "cross-env": "^10.1.0",
-    "eslint": "^9.25.1",
-    "eslint-config-expo": "~10.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint": "^9.38.0",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
     "jest": "^29.2.1",
     "jest-expo": "~54.0.12",
     "lint-staged": "^15.2.7",
     "msw": "^2.11.6",
-    "prettier": "^3.3.2",
     "react-test-renderer": "18.2.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3",
     "webdriverio": "^9.20.0"
-  },
-  "private": true
+  }
 }
