@@ -8,13 +8,14 @@
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "nest build",
-    "prettier": "prettier --write \"src/**/*.ts\"",
+    "prettier": "prettier --config ../../packages/config/prettier.config.json --write \"src/**/*.ts\"",
     "start": "cross-env NODE_ENV=dev nest start",
     "start:dev": "cross-env NODE_ENV=dev nest start --watch",
     "start:debug": "cross-env NODE_ENV=debug nest start --debug --watch",
     "start:prod": "node dist/src/main",
     "start:win": "cross-env NODE_ENV=dev nest start",
-    "lint": "eslint \"/**/*.ts\" --fix",
+    "lint": "eslint . --ignore-pattern '**/.next/**' --ignore-pattern '.history/**'",
+    "lint-fix": "eslint . --ignore-pattern '**/.next/**' --ignore-pattern '.history/**' --fix",
     "pre-commit": "lint-staged",
     "lint-staged": "lint-staged",
     "test": "jest",
@@ -44,19 +45,21 @@
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
+    "@treetracker/config": "0.0.1",
     "@types/express": "^4.17.13",
     "@types/jest": "28.1.8",
     "@types/node": "^16.0.0",
     "@types/supertest": "^6.0.3",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.46.4",
+    "@typescript-eslint/parser": "^8.46.4",
     "cross-env": "^7.0.3",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint": "^9.39.1",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
     "jest": "28.1.3",
     "lint-staged": "^15.2.11",
-    "prettier": "^2.3.2",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
     "ts-jest": "28.0.8",
