@@ -1,11 +1,43 @@
-module.exports = function (api) {
+export default function (api) {
   api.cache(true);
   return {
     presets: [["babel-preset-expo", { unstable_transformImportMeta: true }]],
+    plugins: [
+      [
+        "module-resolver",
+        {
+          root: ["./"],
+          alias: {
+            "@": "./",
+            "@app": "./app",
+            "@assets": "./assets",
+            "@components": "./components",
+            "@common": "./components/ui/common",
+            "@styles": "./styles",
+            "@utils": "./utils",
+            "@hooks": "./hooks",
+            "@constants": "./constants",
+          },
+          extensions: [
+            ".ios.ts",
+            ".android.ts",
+            ".ts",
+            ".ios.tsx",
+            ".android.tsx",
+            ".tsx",
+            ".jsx",
+            ".js",
+            ".json",
+          ],
+        },
+      ],
+      // keep reanimated plugin last
+      "react-native-reanimated/plugin",
+    ],
     env: {
       production: {
-        plugins: ["react-native-paper/babel", "react-native-reanimated/plugin"],
+        plugins: ["react-native-paper/babel"],
       },
     },
   };
-};
+}
