@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   "*.{js,jsx,ts,tsx}": ["eslint --fix --no-warn-ignored", "prettier --write"],
   "*.{json,md,yml,yaml}": ["prettier --write"],
 };
