@@ -7,6 +7,6 @@ it(`renders correctly`, () => {
   const tree = renderer
     .create(<ThemedText>Snapshot test!</ThemedText>)
     .toJSON();
-
+  // @ts-ignore
   expect(tree).toMatchSnapshot();
 });
