@@ -4,7 +4,7 @@ import { User } from "../types/auth";
 export const trimInputs = (input: Partial<User>): User => {
   const trimmed = {} as Record<keyof User, User[keyof User]>;
 
-  (Object.keys(input) as (keyof User)[]).forEach((key) => {
+  (Object.keys(input) as (keyof User)[]).forEach(key => {
     const value = input[key];
     trimmed[key] = typeof value === "string" ? value.trim() : value!;
   });
@@ -16,13 +16,10 @@ export const trimInputs = (input: Partial<User>): User => {
 export const handleValidationError = (
   error: any,
 ): { [key: string]: string } => {
-  return error.details.reduce(
-    (acc: { [key: string]: string }, curr: any) => {
-      if (curr.context?.key) {
-        acc[curr.context.key] = curr.message;
-      }
-      return acc;
-    },
-    {},
-  );
+  return error.details.reduce((acc: { [key: string]: string }, curr: any) => {
+    if (curr.context?.key) {
+      acc[curr.context.key] = curr.message;
+    }
+    return acc;
+  }, {});
 };
