@@ -1,1 +1,2 @@
-export * from "./fetchTokenFromKeycloak";
+export * from "./api/fetchTokenFromKeycloak";
+export * from "./api/deleteAccountFromKeycloak";
