@@ -26,10 +26,17 @@ describe("tests client subscription", () => {
     clientID2 = uuid.v4();
 
     // subscribe clients to a channel, return a promise and verify message/payload from resolved promise
-    subscribe({ pgClient: pgClient, channel: messageObj.channel, clientID: clientID1 }).then(emitter1 => {
-      subscribe({ pgClient: pgClient2, channel: messageObj.channel, clientID: clientID2 }).then(emitter2 => {
-
-        const promise1 = new Promise((resolve) => {
+    subscribe({
+      pgClient: pgClient,
+      channel: messageObj.channel,
+      clientID: clientID1,
+    }).then(emitter1 => {
+      subscribe({
+        pgClient: pgClient2,
+        channel: messageObj.channel,
+        clientID: clientID2,
+      }).then(emitter2 => {
+        const promise1 = new Promise(resolve => {
           emitter1.on("message", message1 => {
             expect(message1).toMatchObject(messageObj); // eslint-disable-line
             expect(new Date(message1.ack[clientID1])).toBeInstanceOf(Date); // eslint-disable-line
@@ -37,7 +44,7 @@ describe("tests client subscription", () => {
           });
         });
 
-        const promise2 = new Promise((resolve) => {
+        const promise2 = new Promise(resolve => {
           emitter2.on("message", message2 => {
             expect(message2).toMatchObject(messageObj); // eslint-disable-line
             expect(new Date(message2.ack[clientID2])).toBeInstanceOf(Date); // eslint-disable-line
@@ -48,9 +55,8 @@ describe("tests client subscription", () => {
         Promise.all([promise1, promise2]).then(() => {
           done();
         });
-
       });
-    })
+    });
 
     // publish message to a given channel
     publish({ pgClient, channel: messageObj.channel, data: messageObj.data });
